+++ index.js
@@ -2,7 +2,6 @@ let banner = require('./banner')
 let chars = require('./chars')
 let fingerprint = require('./fingerprint')
 let getLambdaName = require('./get-lambda-name')
-let getRuntime = require('./get-runtime')
 let pathToUnix = require('./path-to-unix')
 let toLogicalID = require('./to-logical-id')
 let updater = require('./updater')
@@ -12,7 +11,6 @@ module.exports = {
   chars,          // Returns platform appropriate characters for CLI UI printing
   fingerprint,    // Generates public/static.json for `@static fingerprint true`
   getLambdaName,  // Get Lambda name from Arc path
-  getRuntime,     // Get runtime config from Arc file or config
   pathToUnix,     // Normalize to `/` seperated paths for Windows-specific calls
   toLogicalID,    // Converts dash casing into Pascal casing for CloudFormation
   updater,        // Standard Arc status updater and progress indicator